chore(auth): split the Login component
