+++ packages/hedgehog-lab/src/components/Header/Header.tsx
@@ -34,7 +34,7 @@ import {
 import {useCopyToClipboard} from "react-use";
 import {useSnackbar} from "notistack";
 import {ColorModeContext} from '../../App';
-import AuthDialog from "../User/AuthDialog";
+import AuthDialog from "../User/Auth/AuthDialog";
 
 interface HeaderProps {
     siderBarOpen: boolean;
